refactor(settings): split player_settings.rs into per-domain submodules
The 13 enum families colocated in data/src/types/player_settings.rs
each had their own helpers / Display / from_label / as_label / serde
derives, and the PlayerSettings struct that uses them is itself ~200
fields. The 1178-line file made it easy for an AI agent to land a
matching change in only one of the three impl blocks per enum.

Split into 8 per-domain submodules under data/src/types/player_settings/
(library, visualizer, slot_list, navigation, strip, playback, artwork)
with PlayerSettings + sub-module declarations + pub use re-exports in
mod.rs, so all 65 existing crate::types::player_settings::Foo call
sites compile unchanged. The 23 enum tests follow their enums into
the matching sub-file.

No behavioral change. cargo build / clippy --all-targets -D warnings
/ nightly fmt --check / test --workspace all clean (1021 tests pass).
